% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfuncdj.R
\docType{data}
\name{dijkstra}
\alias{dijkstra}
\alias{wiki_graph}
\title{Data Frame for dijkstra}
\format{
An object of class \code{data.frame} with 18 rows and 3 columns.
}
\usage{
wiki_graph
}
\arguments{
\item{v1}{A number.}

\item{v2}{A number.}

\item{w}{A number.}
}
\description{
Data Frame for dijkstra
}
\references{

}
\keyword{datasets}
