% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunc.R
\name{Euclidean}
\alias{Euclidean}
\alias{euclidean}
\title{Euclidean Function}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The greatest common divisor of \code{a} and \code{b}.
}
\description{
Take two numbers a and b, and calculate euclidean distance between them. This basically
calculates greatest common divisor of those numbers. The calculation continues until remainder is 0.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
